tests/requirements.txt
pytest>=7.0
pytest-timeout>=2.1
pyyaml>=6.0
cryptography>=41.0
asn1tools>=0.166
pymongo
